BUG: Accuracy type was using both int and float
Harmonized the types used in the Accuracy type because in one place a float was being used and in the other there was an int.

+++ intelmq/lib/harmonization.py
@@ -289,7 +289,7 @@ class Accuracy(GenericType):
     def sanitize(value):
         try:
             if isinstance(value, bool):
-                return int(value) * 100
+                return float(value) * 100
 
             value = float(value)
             if value >= 0 or value <= 100:
